backend/requirements.txt
# Translation Studio — Python dependencies
# Install with: pip install -r requirements.txt

# --- Web framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12          # file uploads

# --- LLM ---
groq==0.13.0                      # Groq API client (Qwen model)

# --- Embeddings & vector search ---
sentence-transformers==3.3.1
faiss-cpu==1.9.0

# --- NLP ---
spacy==3.8.3
# After install, run: python -m spacy download en_core_web_sm

# --- Spell check ---
pyspellchecker==0.8.1

# --- Document parsing ---
python-docx==1.1.2
PyMuPDF==1.24.14                  # fitz

# --- Utilities ---
pydantic==2.10.3
python-dotenv==1.0.1
numpy==1.26.4
